Use common abstract class to create groups and delete components
This lets the user use either Buttons or SelectMenus to use those
